+++ con_msg.go
@@ -1,0 +1,6 @@
+package pubsub
+
+type PublishedMessage struct {
+	Topic string
+	Data  Binary
+}
